Rooms front-end

+++ database/migrations/2024_10_13_112711_create_reviews_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->foreignId('room_id')->constrained('rooms', 'id')->cascadeOnDelete();
             $table->string('title');
             $table->text('comment')->nullable();
-            $table->decimal('rating', 6, 1);
+            $table->integer('rating');
 
             $table->timestamps();
         });
